V8.2 - nerfed unsus zigzags more! We are happy
